+++ alphadeep/gui.py
@@ -1,5 +1,0 @@
-#!python
-
-
-def run():
-    raise NotImplementedError