+++ misc/stable_pip_install.sh
@@ -1,5 +1,5 @@
-conda create -n alphadeep python=3.8 -y
-conda activate alphadeep
+conda create -n peptdeep python=3.8 -y
+conda activate peptdeep
 pip install -e '../.[stable,development-stable]'
-alphadeep
+peptdeep
 conda deactivate