+++ tests/test_cli.py
@@ -1,11 +1,11 @@
 #!python -m unittest tests.test_utils
-"""This module provides unit tests for alphadeep.cli."""
+"""This module provides unit tests for peptdeep.cli."""
 
 # builtin
 import unittest
 
 # local
-import alphadeep.cli
+import peptdeep.cli
 
 if __name__ == "__main__":
     unittest.main()