+++ tests/test_gui.py
@@ -1,11 +1,11 @@
 #!python -m unittest tests.test_utils
-"""This module provides unit tests for alphadeep.gui."""
+"""This module provides unit tests for peptdeep.gui."""
 
 # builtin
 import unittest
 
 # local
-import alphadeep.gui
+import peptdeep.gui
 
 if __name__ == "__main__":
     unittest.main()